Implement main menu and game screen
Abstract away some of the initial structure from the tutorial
