+++ src/main/java/biblioteket/roborally/game/IGame.java
@@ -1,0 +1,5 @@
+package biblioteket.roborally.game;
+
+public interface IGame {
+
+}
