+++ src/main/java/biblioteket/roborally/game/MainMenuScreen.java
@@ -1,0 +1,71 @@
+package biblioteket.roborally.game;
+
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Screen;
+import com.badlogic.gdx.graphics.GL20;
+import com.badlogic.gdx.graphics.OrthographicCamera;
+
+/**
+ * The landing screen once the user starts the game, this is where the player(s)
+ * select how many players there are, multiplayer etc. Currently only a simple
+ * canvas is displayed that the user can click and then the game launches.
+ */
+public class MainMenuScreen implements Screen {
+    private final RoboRally game;
+    OrthographicCamera camera;
+
+    public MainMenuScreen(final RoboRally game) {
+        this.game = game;
+        this.camera = new OrthographicCamera();
+        camera.setToOrtho(false, 640, 640);
+    }
+
+    @Override
+    public void show() {
+
+    }
+
+    @Override
+    public void render(float delta) {
+        Gdx.gl.glClearColor(0, 0, 0.2f, 1);
+        Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
+
+        camera.update();
+        game.getBatch().setProjectionMatrix(camera.combined);
+
+        game.getBatch().begin();
+        game.getFont().draw(game.getBatch(), "ROBORALLY", 300, 320);
+        game.getFont().draw(game.getBatch(), "Tap anywhere to begin!", 300, 300);
+        game.getBatch().end();
+
+        if (Gdx.input.isTouched()) {
+            game.setScreen(new GameScreen(game));
+            dispose();
+        }
+    }
+
+    @Override
+    public void resize(int width, int height) {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void hide() {
+
+    }
+
+    @Override
+    public void dispose() {
+
+    }
+}
